add right and left arrows to carousel header, add functionality so that on any element click in the header, the overlay is rendered, reorganize client folder
