+++ src/client/overlay-module/leftArrow.jsx
