+++ src/client/overlay-module/rightArrow.jsx
